refacting: engine add Empty & new, move Formats 2 rules & adjust tests accordingly
